enable install options in package resource
I recently had a case where I wanted to pass additional install options to yum on a centos install of tomcat 7.  The particular options were to disable a project repo which we had provisioned which also contained a tomcat package but at version 6.  Without the ability to pass '--disablerepo=myrepo' to the effective yum command, I could not get the module install to install the EPEL tomcat version, in favour of our own package.

To facilitate passing extra install options I have added this attribute to the package resource.