Specify port when purging connectors
This commit change the behavior of 'purge_connectors' so that it will
purge ports of the same protocol that have a different port than
specified.

Prior to this commit, *all* connectors that matched the protocol were
purged, including the desired one.  This resulted in a purge/create
every Puppet run.

Refer to https://tickets.puppetlabs.com/browse/MODULES-2471

Update spec for purge_connectors

Should purge connectors that match the protocol and have a different
port

Provide more clarity for purge_connectors in readme