(#2984) Add support for host aliases
Add an optional $aliases parameter to tomcat::config::server::host
to allow the management of Alias entries within the Host entry.
If the parameter is omitted, any existing Alias entries within the Host
XML element are left unaltered.  Otherwise, the set of Alias entries
within the Host entry are set to match the content of the $aliases array.
(An empty array is supported, which allows any existing Alias entries to be
removed.)