Fix base/home installation
Tomcat is traditionally installed into catalina_home, and multiple
instances can be created by using a custom catalina_base per instance.
The previous iteration of this was buggy, so this commit implements a
new defined resource tomcat::install which is used for declaring each
install of the software into a catalina_home, and then tomcat::instance
can deploy an instance from catalina_home to catalina_base and declare a
tomcat::service to go along with it.

The base class is used to set a few defaults to be used by all
instances.

If the old behavior is to be used, this is still possible.