(MODULES-3117) User and Group bugfix to support setting via tomcat::instance and starting the service.
Previously setting the user and group can only be done once via the tomcat class. It appears there has been some work to support setting the user and group in tomcat::instance but some other changes had to be made so the service would start up correctly with the user set in the instance.

The change defaults to the user and group set in the tomcat class, but can use the tomcat instance variables if those have been set.