(MODULES-4658) Allow duplicate realms
Without this PR, the className of a realm may not be unique (eg, in the
case of two ldap authentication realms both using className of
org.apache.catalina.realm.JNDIRealm) and thus augeas happily overwrites
the first realm with a second every puppet run.

This adds an extra attribute puppetName based on the resource title to
maintain a uniqueness constraint, and will look for realms with a given
className (the old case) as well as realms with the puppetName. There
can be no more than one unique className managed by puppet currently, so
this handles upgrades by adding a puppetName to those existing realms,
and allowing duplicates after that point.