(maint) Don't duplicate resources for different dependencies
The file resource is declared with ensure_resource() which does not
cause duplicate resource declarations unless some of the attributes are
different from the already-declared resource.

In the case of this dependency, it will be different for each
declaration of tomcat_user so must be defined separately from the
ensure_resource call.