(maint) Deprecate install_from_source in the base class
The logic for install_from_source was removed from the base class and is
in the tomcat::install exclusively in 328a97d . The reason for this is
that the value must be changed on exactly one installation of tomcat due
to the fact that it installs a package. Having a "global default" for
it does not make sense.