Completely remove tomcat::install_from_source
This parameter is not used anymore and was deprecated more than 2 years
ago.  2 major versions of the module where released with this
deprecation.

When set, this parameter makes the catalog compilation fail with an
error message explaining that this parameter is deprecated.  However,
when set to false or undef, no error is raised, but the default
behavior is still to install from source…

Completely remove this parameter so that setting it to any value will
produce an error.