(CONT-802) - Revert RSpec/NoExpectationExample
During the work to add support for puppet 8, a rubocop violation
"RSpec/NoExpectationExample" was fixed.

This included adding an expect {} block around puppet applys, which
initially seemed to work as expected.

However it was discovered that this hid error outputs of failed runs,
meaning debugging was impossible.
This commit reverts these changes and adds the cop back to the "todo".